refactor(agents): rename skills wrapper module to skill_wrapper
